+++ src/plonemeeting/restapi/services/add.py
@@ -2,6 +2,7 @@
 
 from collective.iconifiedcategory.utils import calculate_category_id
 from imio.helpers.content import get_vocab
+from imio.helpers.content import uuidToObject
 from imio.helpers.security import fplog
 from imio.history.utils import add_event_to_wf_history
 from imio.restapi.services.add import FolderPost
@@ -10,7 +11,6 @@ from plone import api
 from plonemeeting.restapi.config import CONFIG_ID_ERROR
 from plonemeeting.restapi.config import CONFIG_ID_NOT_FOUND_ERROR
 from plonemeeting.restapi.utils import check_in_name_of
-from plonemeeting.restapi.utils import rest_uuid_to_object
 from Products.PloneMeeting.utils import get_annexes_config
 from Products.PloneMeeting.utils import org_id_to_uid
 from zExceptions import BadRequest
@@ -329,7 +329,7 @@ class AnnexPost(BasePost):
     @property
     def container(self):
         if not self._container:
-            self._container = rest_uuid_to_object(self.container_uid)
+            self._container = uuidToObject(self.container_uid)
             self.context = self._container
         return self.context